+++ creme/compose/__init__.py
@@ -2,7 +2,14 @@
 Meta-estimators for building composite models.
 """
 from .pipeline import Pipeline
+from .target_modifier import TargetModifierRegressor
+from .target_modifier import BoxCoxTransformRegressor
 from .union import TransformerUnion
 
 
-__all__ = ['Pipeline', 'TransformerUnion']
+__all__ = [
+    'BoxCoxTransformRegressor',
+    'Pipeline',
+    'TargetModifierRegressor',
+    'TransformerUnion'
+]