+++ creme/stats/__init__.py
@@ -12,7 +12,14 @@ from .mode import Mode
 from .n_unique import NUnique
 from .ptp import PeakToPeak
 from .quantile import Quantile
-from .quantile import RollingQuantile
+from .rolling_max import RollingMax
+from .rolling_mean import RollingMean
+from .rolling_min import RollingMin
+from .rolling_mode import RollingMode
+from .rolling_ptp import RollingPeakToPeak
+from .rolling_quantile import RollingQuantile
+from .rolling_sum import RollingSum
+from .rolling_variance import RollingVariance
 from .skew import Skew
 from .sum import Sum
 from .variance import Variance
@@ -30,7 +37,14 @@ __all__ = [
     'NUnique',
     'PeakToPeak',
     'Quantile',
+    'RollingMean',
     'RollingQuantile',
+    'RollingMax',
+    'RollingMin',
+    'RollingMode',
+    'RollingPeakToPeak',
+    'RollingSum',
+    'RollingVariance',
     'Skew',
     'Sum',
     'Variance'