+++ creme/stats/rolling_max.py
@@ -1,0 +1,39 @@
+from . import base
+from . import _window
+
+
+class RollingMax(base.RunningStatistic, _window.SortedWindow):
+    """Computes a windowed running max.
+
+    Attributes:
+        window_size (int): Size of the rolling window.
+
+    Example:
+
+    ::
+
+        >>> from creme import stats
+
+        >>> X = [1, -4, 3, -2, 2, 1]
+        >>> rolling_max = stats.RollingMax(window_size=2)
+        >>> for x in X:
+        ...     print(rolling_max.update(x).get())
+        1
+        1
+        3
+        3
+        2
+        2
+
+    """
+
+    @property
+    def name(self):
+        return 'rolling_{self.window_size}_max'
+
+    def update(self, x):
+        self.append(x)
+        return self
+
+    def get(self):
+        return self[-1]